refactor header symbols. seems to break the the macro system.
