+++ include/El/blas_like/level3.hpp
@@ -9,621 +9,25 @@
 #ifndef EL_BLAS3_HPP
 #define EL_BLAS3_HPP
 
-namespace El {
-
-// Gemm
-// ====
-namespace GemmAlgorithmNS {
-enum GemmAlgorithm {
-  GEMM_DEFAULT,
-  GEMM_SUMMA_A,
-  GEMM_SUMMA_B,
-  GEMM_SUMMA_C,
-  GEMM_SUMMA_DOT,
-  GEMM_CANNON
-};
-}
-using namespace GemmAlgorithmNS;
-
-template<typename T>
-void Gemm
-( Orientation orientA, Orientation orientB,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B, T beta, Matrix<T>& C );
-
-template<typename T>
-void Gemm
-( Orientation orientA, Orientation orientB,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B, Matrix<T>& C );
-
-template<typename T>
-void Gemm
-( Orientation orientA, Orientation orientB,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-  T beta,        ElementalMatrix<T>& C, GemmAlgorithm alg=GEMM_DEFAULT );
-
-template<typename T>
-void Gemm
-( Orientation orientA, Orientation orientB,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-                 ElementalMatrix<T>& C, GemmAlgorithm alg=GEMM_DEFAULT );
-
-template<typename T>
-void LocalGemm
-( Orientation orientA, Orientation orientB,
-  T alpha, const ElementalMatrix<T>& A,
-           const ElementalMatrix<T>& B,
-  T beta,        ElementalMatrix<T>& C );
-template<typename T>
-void LocalGemm
-( Orientation orientA, Orientation orientB,
-  T alpha, const ElementalMatrix<T>& A,
-           const ElementalMatrix<T>& B,
-                 ElementalMatrix<T>& C );
-
-// Hemm
-// ====
-template<typename T>
-void Hemm
-( LeftOrRight side, UpperOrLower uplo,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B, T beta, Matrix<T>& C );
-
-template<typename T>
-void Hemm
-( LeftOrRight side, UpperOrLower uplo,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-  T beta,        ElementalMatrix<T>& C );
-
-// Herk
-// ====
-template<typename T>
-void Herk
-( UpperOrLower uplo, Orientation orientation,
-  Base<T> alpha, const Matrix<T>& A, Base<T> beta, Matrix<T>& C );
-template<typename T>
-void Herk
-( UpperOrLower uplo, Orientation orientation,
-  Base<T> alpha, const Matrix<T>& A, Matrix<T>& C );
-
-template<typename T>
-void Herk
-( UpperOrLower uplo, Orientation orientation,
-  Base<T> alpha, const ElementalMatrix<T>& A, 
-  Base<T> beta,        ElementalMatrix<T>& C );
-template<typename T>
-void Herk
-( UpperOrLower uplo, Orientation orientation,
-  Base<T> alpha, const ElementalMatrix<T>& A, ElementalMatrix<T>& C );
-
-template<typename T>
-void Herk
-( UpperOrLower uplo, Orientation orientation,
-  Base<T> alpha, const SparseMatrix<T>& A, 
-  Base<T> beta,        SparseMatrix<T>& C );
-template<typename T>
-void Herk
-( UpperOrLower uplo, Orientation orientation,
-  Base<T> alpha, const SparseMatrix<T>& A, 
-                       SparseMatrix<T>& C );
-
-template<typename T>
-void Herk
-( UpperOrLower uplo, Orientation orientation,
-  Base<T> alpha, const DistSparseMatrix<T>& A,
-  Base<T> beta,        DistSparseMatrix<T>& C );
-template<typename T> 
-void Herk
-( UpperOrLower uplo, Orientation orientation,
-  Base<T> alpha, const DistSparseMatrix<T>& A,
-                       DistSparseMatrix<T>& C );
-
-// Her2k
-// =====
-template<typename T>
-void Her2k
-( UpperOrLower uplo, Orientation orientation,
-  T alpha,       const Matrix<T>& A, const Matrix<T>& B, 
-  Base<T> beta,        Matrix<T>& C );
-
-template<typename T>
-void Her2k
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B, Matrix<T>& C );
-
-template<typename T>
-void Her2k
-( UpperOrLower uplo, Orientation orientation,
-  T alpha,      const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-  Base<T> beta,       ElementalMatrix<T>& C );
-
-template<typename T>
-void Her2k
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-                 ElementalMatrix<T>& C );
-
-// Multiply
-// ========
-// NOTE: The following routine multiplies a sparse matrix by a set of vectors
-//       and is obviously not a BLAS routine. However, it is a basic linear
-//       algebra routine making use of Elemental's core data structures, and
-//       so this is the natural placement
-template<typename T>
-void Multiply
-( Orientation orientation,
-  T alpha, const SparseMatrix<T>& A, const Matrix<T>& X,
-  T beta,                                  Matrix<T>& Y );
-
-template<typename T>
-void Multiply
-( Orientation orientation,
-  T alpha,
-  const DistSparseMatrix<T>& A,
-  const DistMultiVec<T>& X,
-  T beta,
-        DistMultiVec<T>& Y );
-template<typename T>
-void Multiply
-( Orientation orientation,
-  T alpha,
-  const DistSparseMatrix<T>& A,
-  const AbstractDistMatrix<T>& X,
-  T beta,
-        AbstractDistMatrix<T>& Y );
-
-// MultiShiftQuasiTrsm
-// ===================
-template<typename F>
-void MultiShiftQuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  F alpha, const Matrix<F>& A, const Matrix<F>& shifts, Matrix<F>& B );
-template<typename Real>
-void MultiShiftQuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  Complex<Real> alpha,
-  const Matrix<Real>& A,
-  const Matrix<Complex<Real>>& shifts,
-        Matrix<Real>& BReal, Matrix<Real>& BImag );
-template<typename F>
-void MultiShiftQuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  F alpha, const ElementalMatrix<F>& A, const ElementalMatrix<F>& shifts,
-  ElementalMatrix<F>& B );
-template<typename Real>
-void MultiShiftQuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  Complex<Real> alpha,
-  const ElementalMatrix<Real>& A,
-  const ElementalMatrix<Complex<Real>>& shifts,
-        ElementalMatrix<Real>& BReal, ElementalMatrix<Real>& BImag );
-
-template<typename F>
-void LocalMultiShiftQuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  F alpha, const DistMatrix<F,STAR,STAR>& A,
-           const ElementalMatrix<F>& shifts,
-                 ElementalMatrix<F>& X );
-template<typename Real>
-void LocalMultiShiftQuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  Complex<Real> alpha,
-  const DistMatrix<Real,STAR,STAR>& A,
-  const ElementalMatrix<Complex<Real>>& shifts,
-        ElementalMatrix<Real>& XReal,
-        ElementalMatrix<Real>& XImag );
-
-// MultiShiftTrsm
-// ==============
-template<typename F>
-void MultiShiftTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  F alpha, Matrix<F>& U, const Matrix<F>& shifts, Matrix<F>& X );
-template<typename F>
-void MultiShiftTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  F alpha, const ElementalMatrix<F>& U, const ElementalMatrix<F>& shifts,
-  ElementalMatrix<F>& X );
-
-// QuasiTrsm
-// =========
-template<typename F>
-void QuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  F alpha, const Matrix<F>& A, Matrix<F>& B,
-  bool checkIfSingular=false );
-template<typename F>
-void QuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  F alpha, const ElementalMatrix<F>& A, ElementalMatrix<F>& B,
-  bool checkIfSingular=false );
-
-template<typename F>
-void LocalQuasiTrsm
-( LeftOrRight side, UpperOrLower uplo, Orientation orientation,
-  F alpha, const DistMatrix<F,STAR,STAR>& A, ElementalMatrix<F>& X,
-  bool checkIfSingular=false );
-
-// Symm
-// ====
-template<typename T>
-void Symm
-( LeftOrRight side, UpperOrLower uplo,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B, T beta, Matrix<T>& C,
-  bool conjugate=false );
-template<typename T>
-void Symm
-( LeftOrRight side, UpperOrLower uplo,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-  T beta,        ElementalMatrix<T>& C, bool conjugate=false );
-
-namespace symm {
-
-template<typename T>
-void LocalAccumulateLL
-( Orientation orientation, T alpha,
-  const DistMatrix<T>& A,
-  const DistMatrix<T,MC,  STAR>& B_MC_STAR,
-  const DistMatrix<T,STAR,MR  >& BTrans_STAR_MR,
-        DistMatrix<T,MC,  STAR>& Z_MC_STAR,
-        DistMatrix<T,MR,  STAR>& Z_MR_STAR );
-
-template<typename T>
-void LocalAccumulateLU
-( Orientation orientation, T alpha,
-  const DistMatrix<T>& A,
-  const DistMatrix<T,MC,  STAR>& B_MC_STAR,
-  const DistMatrix<T,STAR,MR  >& BTrans_STAR_MR,
-        DistMatrix<T,MC,  STAR>& Z_MC_STAR,
-        DistMatrix<T,MR,  STAR>& Z_MR_STAR );
-
-template<typename T>
-void LocalAccumulateRL
-( Orientation orientation, T alpha,
-  const DistMatrix<T>& A,
-  const DistMatrix<T,STAR,MC  >& B_STAR_MC,
-  const DistMatrix<T,MR,  STAR>& BTrans_MR_STAR,
-        DistMatrix<T,MC,  STAR>& ZTrans_MC_STAR,
-        DistMatrix<T,MR,  STAR>& ZTrans_MR_STAR );
-
-template<typename T>
-void LocalAccumulateRU
-( Orientation orientation, T alpha,
-  const DistMatrix<T,MC,  MR  >& A,
-  const DistMatrix<T,STAR,MC  >& B_STAR_MC,
-  const DistMatrix<T,MR,  STAR>& BTrans_MR_STAR,
-        DistMatrix<T,MC,  STAR>& ZTrans_MC_STAR,
-        DistMatrix<T,MR,  STAR>& ZTrans_MR_STAR );
-
-} // namespace symm
-
-// Syrk
-// ====
-template<typename T>
-void Syrk
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const Matrix<T>& A, T beta, Matrix<T>& C,
-  bool conjugate=false );
-template<typename T>
-void Syrk
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const Matrix<T>& A, Matrix<T>& C,
-  bool conjugate=false );
-
-template<typename T>
-void Syrk
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const ElementalMatrix<T>& A, 
-  T beta,        ElementalMatrix<T>& C, bool conjugate=false );
-template<typename T>
-void Syrk
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const ElementalMatrix<T>& A, ElementalMatrix<T>& C,
-  bool conjugate=false );
-
-template<typename T>
-void Syrk
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const SparseMatrix<T>& A, 
-  T beta,        SparseMatrix<T>& C, bool conjugate=false );
-template<typename T>
-void Syrk
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const SparseMatrix<T>& A, 
-                 SparseMatrix<T>& C, bool conjugate=false );
-
-template<typename T>
-void Syrk
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const DistSparseMatrix<T>& A, 
-  T beta,        DistSparseMatrix<T>& C, bool conjugate=false );
-template<typename T>
-void Syrk
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const DistSparseMatrix<T>& A, 
-                 DistSparseMatrix<T>& C, bool conjugate=false );
-
-// Syr2k
-// =====
-template<typename T>
-void Syr2k
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B, T beta, Matrix<T>& C,
-  bool conjugate=false );
-
-template<typename T>
-void Syr2k
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B, Matrix<T>& C,
-  bool conjugate=false );
-
-template<typename T>
-void Syr2k
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-  T beta,        ElementalMatrix<T>& C,
-  bool conjugate=false );
-
-template<typename T>
-void Syr2k
-( UpperOrLower uplo, Orientation orientation,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-                 ElementalMatrix<T>& C,
-  bool conjugate=false );
-
-// Trdtrmm
-// =======
-template<typename F>
-void Trdtrmm( UpperOrLower uplo, Matrix<F>& A, bool conjugate=false );
-template<typename F>
-void Trdtrmm
-( UpperOrLower uplo, Matrix<F>& A, const Matrix<F>& dOff, 
-  bool conjugate=false );
-
-template<typename F>
-void Trdtrmm
-( UpperOrLower uplo, ElementalMatrix<F>& A, bool conjugate=false );
-template<typename F>
-void Trdtrmm
-( UpperOrLower uplo,
-  ElementalMatrix<F>& A, const ElementalMatrix<F>& dOff, 
-  bool conjugate=false );
-
-template<typename F>
-void Trdtrmm
-( UpperOrLower uplo, DistMatrix<F,STAR,STAR>& A, bool conjugate=false );
-template<typename F>
-void Trdtrmm
-( UpperOrLower uplo,
-  DistMatrix<F,STAR,STAR>& A, const DistMatrix<F,STAR,STAR>& dOff,
-  bool conjugate=false );
-
-// Trmm
-// ====
-template<typename T>
-void Trmm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  T alpha, const Matrix<T>& A, Matrix<T>& B );
-template<typename T>
-void Trmm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  T alpha, const ElementalMatrix<T>& A, ElementalMatrix<T>& X );
-
-template<typename T>
-void LocalTrmm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  T alpha, const DistMatrix<T,STAR,STAR>& A, ElementalMatrix<T>& B );
-
-// Trsm
-// ====
-namespace TrsmAlgorithmNS {
-enum TrsmAlgorithm {
-  TRSM_DEFAULT,
-  TRSM_LARGE,
-  TRSM_MEDIUM,
-  TRSM_SMALL
-};
-}
-using namespace TrsmAlgorithmNS;
-
-template<typename F>
-void Trsm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  F alpha, const Matrix<F>& A, Matrix<F>& B,
-  bool checkIfSingular=false );
-template<typename F>
-void Trsm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  F alpha,
-  const AbstractDistMatrix<F>& A,
-        AbstractDistMatrix<F>& B,
-  bool checkIfSingular=false, TrsmAlgorithm alg=TRSM_DEFAULT );
-
-template<typename F>
-void LocalTrsm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  F alpha,
-  const DistMatrix<F,STAR,STAR>& A,
-        AbstractDistMatrix<F>& X,
-  bool checkIfSingular=false );
-
-// Trstrm
-// ======
-template<typename F>
-void Trstrm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  F alpha, const Matrix<F>& A, Matrix<F>& X,
-  bool checkIfSingular=true );
-template<typename F>
-void Trstrm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  F alpha, const ElementalMatrix<F>& A, ElementalMatrix<F>& X,
-  bool checkIfSingular=true );
-template<typename F>
-void Trstrm
-( LeftOrRight side, UpperOrLower uplo,
-  Orientation orientation, UnitOrNonUnit diag,
-  F alpha, const DistMatrix<F,STAR,STAR>& A, DistMatrix<F,STAR,STAR>& X,
-  bool checkIfSingular=true );
-
-// Trtrmm
-// ======
-template<typename T>
-void Trtrmm( UpperOrLower uplo, Matrix<T>& A, bool conjugate=false );
-template<typename T>
-void Trtrmm
-( UpperOrLower uplo, ElementalMatrix<T>& A, bool conjugate=false );
-template<typename T>
-void Trtrmm
-( UpperOrLower uplo, DistMatrix<T,STAR,STAR>& A, bool conjugate=false );
-
-// TwoSidedTrmm
-// ============
-template<typename T>
-void TwoSidedTrmm
-( UpperOrLower uplo, UnitOrNonUnit diag, 
-  Matrix<T>& A, const Matrix<T>& B );
-template<typename T>
-void TwoSidedTrmm
-( UpperOrLower uplo, UnitOrNonUnit diag, 
-  ElementalMatrix<T>& A, const ElementalMatrix<T>& B );
-template<typename T>
-void TwoSidedTrmm
-( UpperOrLower uplo, UnitOrNonUnit diag,
-  DistMatrix<T,MC,MR,BLOCK>& A, const DistMatrix<T,MC,MR,BLOCK>& B );
-template<typename T>
-void LocalTwoSidedTrmm
-( UpperOrLower uplo, UnitOrNonUnit diag,
-  DistMatrix<T,STAR,STAR>& A, const DistMatrix<T,STAR,STAR>& B );
-
-// TwoSidedTrsm
-// ============
-template<typename F>
-void TwoSidedTrsm
-( UpperOrLower uplo, UnitOrNonUnit diag, 
-  Matrix<F>& A, const Matrix<F>& B );
-template<typename F>
-void TwoSidedTrsm
-( UpperOrLower uplo, UnitOrNonUnit diag,
-  ElementalMatrix<F>& A, const ElementalMatrix<F>& B );
-template<typename F>
-void TwoSidedTrsm
-( UpperOrLower uplo, UnitOrNonUnit diag,
-  DistMatrix<F,MC,MR,BLOCK>& A, const DistMatrix<F,MC,MR,BLOCK>& B );
-template<typename F>
-void TwoSidedTrsm
-( UpperOrLower uplo, UnitOrNonUnit diag,
-  DistMatrix<F,STAR,STAR>& A, const DistMatrix<F,STAR,STAR>& B );
-
-// Trrk
-// ====
-template<typename T>
-void Trrk
-( UpperOrLower uplo, 
-  Orientation orientA, Orientation orientB,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B,
-  T beta,        Matrix<T>& C );
-template<typename T>
-void Trrk
-( UpperOrLower uplo, 
-  Orientation orientA, Orientation orientB,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-  T beta,        ElementalMatrix<T>& C );
-template<typename T>
-void LocalTrrk
-( UpperOrLower uplo,
-  T alpha, const DistMatrix<T,MC,  STAR>& A,
-           const DistMatrix<T,STAR,MR  >& B,
-  T beta,        DistMatrix<T,MC,  MR  >& C );
-template<typename T>
-void LocalTrrk
-( UpperOrLower uplo,
-  Orientation orientB,
-  T alpha, const DistMatrix<T,MC,STAR>& A,
-           const DistMatrix<T,MR,STAR>& B,
-  T beta,        DistMatrix<T>& C );
-template<typename T>
-void LocalTrrk
-( UpperOrLower uplo,
-  Orientation orientA,
-  T alpha, const DistMatrix<T,STAR,MC>& A,
-           const DistMatrix<T,STAR,MR>& B,
-  T beta,        DistMatrix<T,MC,  MR>& C );
-template<typename T>
-void LocalTrrk
-( UpperOrLower uplo,
-  Orientation orientA, Orientation orientB,
-  T alpha, const DistMatrix<T,STAR,MC  >& A,
-           const DistMatrix<T,MR,  STAR>& B,
-  T beta,        DistMatrix<T,MC,  MR  >& C );
-
-// Trr2k
-// =====
-/*
-template<typename T>
-void Trr2k
-( UpperOrLower uplo, 
-  Orientation orientA, Orientation orientB,
-  Orientation orientC, Orientation orientD,
-  T alpha, const Matrix<T>& A, const Matrix<T>& B,
-  T beta,  const Matrix<T>& C, const Matrix<T>& D,
-  Tgamma,        Matrix<T>& E );
-*/
-template<typename T>
-void Trr2k
-( UpperOrLower uplo,
-  Orientation orientA, Orientation orientB,
-  Orientation orientC, Orientation orientD,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-  T beta,  const ElementalMatrix<T>& C, const ElementalMatrix<T>& D,
-  T gamma,       ElementalMatrix<T>& E );
-
-// The distributions of the oriented matrices must match
-template<typename T>
-void LocalTrr2k
-( UpperOrLower uplo,
-  Orientation orientA, Orientation orientB,
-  Orientation orientC, Orientation orientD,
-  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
-  T beta,  const ElementalMatrix<T>& C, const ElementalMatrix<T>& D,
-  T gamma,       ElementalMatrix<T>& E );
-
-// Hermitian from EVD
-// ==================
-// A := Z diag(w) Z^H, where w is real
-template<typename F>
-void HermitianFromEVD
-( UpperOrLower uplo,
-        Matrix<F>& A,
-  const Matrix<Base<F>>& w,
-  const Matrix<F>& Z );
-template<typename F>
-void HermitianFromEVD
-( UpperOrLower uplo,
-        ElementalMatrix<F>& A,
-  const ElementalMatrix<Base<F>>& w,
-  const ElementalMatrix<F>& Z );
-
-// Normal from EVD
-// ===============
-// A := Z diag(w) Z^H, where w is complex
-template<typename Real>
-void NormalFromEVD
-(       Matrix<Complex<Real>>& A,
-  const Matrix<Complex<Real>>& w,
-  const Matrix<Complex<Real>>& Z );
-template<typename Real>
-void NormalFromEVD
-(       ElementalMatrix<Complex<Real>>& A,
-  const ElementalMatrix<Complex<Real>>& w,
-  const ElementalMatrix<Complex<Real>>& Z );
-
-} // namespace El
+#include <El/blas_like/level3/Gemm.hpp>
+#include <El/blas_like/level3/Hemm.hpp>
+#include <El/blas_like/level3/Her2k.hpp>
+#include <El/blas_like/level3/Herk.hpp>
+#include <El/blas_like/level3/MultiShiftQuasiTrsm.hpp>
+#include <El/blas_like/level3/MultiShiftTrsm.hpp>
+#include <El/blas_like/level3/Multiply.hpp>
+#include <El/blas_like/level3/QuasiTrsm.hpp>
+#include <El/blas_like/level3/Symm.hpp>
+#include <El/blas_like/level3/Syr2k.hpp>
+#include <El/blas_like/level3/Syrk.hpp>
+#include <El/blas_like/level3/Trdtrmm.hpp>
+#include <El/blas_like/level3/Trmm.hpp>
+#include <El/blas_like/level3/Trr2k.hpp>
+#include <El/blas_like/level3/Trrk.hpp>
+#include <El/blas_like/level3/Trsm.hpp>
+#include <El/blas_like/level3/Trstrm.hpp>
+#include <El/blas_like/level3/Trtrmm.hpp>
+#include <El/blas_like/level3/TwoSidedTrmm.hpp>
+#include <El/blas_like/level3/TwoSidedTrsm.hpp>
 
 #endif // ifndef EL_BLAS3_HPP
