+++ include/El/blas_like/level3/Hemm.hpp
@@ -1,0 +1,29 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_HEMM_HPP
+#define EL_BLAS3_HEMM_HPP
+
+namespace El {
+
+// Hemm
+// ====
+template<typename T>
+void Hemm
+( LeftOrRight side, UpperOrLower uplo,
+  T alpha, const Matrix<T>& A, const Matrix<T>& B, T beta, Matrix<T>& C );
+
+template<typename T>
+void Hemm
+( LeftOrRight side, UpperOrLower uplo,
+  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
+  T beta,        ElementalMatrix<T>& C );
+
+} //end namespace El
+
+#endif //EL_BLAS3_HEMM_HPP
