+++ include/El/blas_like/level3/Herk.hpp
@@ -1,0 +1,59 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_HERK_HPP
+#define EL_BLAS3_HERK_HPP
+
+namespace El {
+
+// Herk
+// ====
+template<typename T>
+void Herk
+( UpperOrLower uplo, Orientation orientation,
+  Base<T> alpha, const Matrix<T>& A, Base<T> beta, Matrix<T>& C );
+template<typename T>
+void Herk
+( UpperOrLower uplo, Orientation orientation,
+  Base<T> alpha, const Matrix<T>& A, Matrix<T>& C );
+
+template<typename T>
+void Herk
+( UpperOrLower uplo, Orientation orientation,
+  Base<T> alpha, const ElementalMatrix<T>& A, 
+  Base<T> beta,        ElementalMatrix<T>& C );
+template<typename T>
+void Herk
+( UpperOrLower uplo, Orientation orientation,
+  Base<T> alpha, const ElementalMatrix<T>& A, ElementalMatrix<T>& C );
+
+template<typename T>
+void Herk
+( UpperOrLower uplo, Orientation orientation,
+  Base<T> alpha, const SparseMatrix<T>& A, 
+  Base<T> beta,        SparseMatrix<T>& C );
+template<typename T>
+void Herk
+( UpperOrLower uplo, Orientation orientation,
+  Base<T> alpha, const SparseMatrix<T>& A, 
+                       SparseMatrix<T>& C );
+
+template<typename T>
+void Herk
+( UpperOrLower uplo, Orientation orientation,
+  Base<T> alpha, const DistSparseMatrix<T>& A,
+  Base<T> beta,        DistSparseMatrix<T>& C );
+template<typename T> 
+void Herk
+( UpperOrLower uplo, Orientation orientation,
+  Base<T> alpha, const DistSparseMatrix<T>& A,
+                       DistSparseMatrix<T>& C );
+
+} //end namespace El
+
+#endif
