+++ include/El/blas_like/level3/Symm.hpp
@@ -1,0 +1,68 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_SYMM_HPP
+#define EL_BLAS3_SYMM_HPP
+
+namespace El {
+
+// Symm
+// ====
+template<typename T>
+void Symm
+( LeftOrRight side, UpperOrLower uplo,
+  T alpha, const Matrix<T>& A, const Matrix<T>& B, T beta, Matrix<T>& C,
+  bool conjugate=false );
+template<typename T>
+void Symm
+( LeftOrRight side, UpperOrLower uplo,
+  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
+  T beta,        ElementalMatrix<T>& C, bool conjugate=false );
+
+namespace symm {
+
+template<typename T>
+void LocalAccumulateLL
+( Orientation orientation, T alpha,
+  const DistMatrix<T>& A,
+  const DistMatrix<T,MC,  STAR>& B_MC_STAR,
+  const DistMatrix<T,STAR,MR  >& BTrans_STAR_MR,
+        DistMatrix<T,MC,  STAR>& Z_MC_STAR,
+        DistMatrix<T,MR,  STAR>& Z_MR_STAR );
+
+template<typename T>
+void LocalAccumulateLU
+( Orientation orientation, T alpha,
+  const DistMatrix<T>& A,
+  const DistMatrix<T,MC,  STAR>& B_MC_STAR,
+  const DistMatrix<T,STAR,MR  >& BTrans_STAR_MR,
+        DistMatrix<T,MC,  STAR>& Z_MC_STAR,
+        DistMatrix<T,MR,  STAR>& Z_MR_STAR );
+
+template<typename T>
+void LocalAccumulateRL
+( Orientation orientation, T alpha,
+  const DistMatrix<T>& A,
+  const DistMatrix<T,STAR,MC  >& B_STAR_MC,
+  const DistMatrix<T,MR,  STAR>& BTrans_MR_STAR,
+        DistMatrix<T,MC,  STAR>& ZTrans_MC_STAR,
+        DistMatrix<T,MR,  STAR>& ZTrans_MR_STAR );
+
+template<typename T>
+void LocalAccumulateRU
+( Orientation orientation, T alpha,
+  const DistMatrix<T,MC,  MR  >& A,
+  const DistMatrix<T,STAR,MC  >& B_STAR_MC,
+  const DistMatrix<T,MR,  STAR>& BTrans_MR_STAR,
+        DistMatrix<T,MC,  STAR>& ZTrans_MC_STAR,
+        DistMatrix<T,MR,  STAR>& ZTrans_MR_STAR );
+
+} //end namespace symm
+} //end namespace ayasdi
+
+#endif //EL_BLAS3_SYMM_HPP
