+++ include/El/blas_like/level3/Syr2k.hpp
@@ -1,0 +1,44 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_SYR2K_HPP
+#define EL_BLAS3_SYR2K_HPP
+
+namespace El {
+
+// Syr2k
+// =====
+template<typename T>
+void Syr2k
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const Matrix<T>& A, const Matrix<T>& B, T beta, Matrix<T>& C,
+  bool conjugate=false );
+
+template<typename T>
+void Syr2k
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const Matrix<T>& A, const Matrix<T>& B, Matrix<T>& C,
+  bool conjugate=false );
+
+template<typename T>
+void Syr2k
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
+  T beta,        ElementalMatrix<T>& C,
+  bool conjugate=false );
+
+template<typename T>
+void Syr2k
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
+                 ElementalMatrix<T>& C,
+  bool conjugate=false );
+
+} // namespace El
+
+#endif // ifndef EL_BLAS3_SYR2K_HPP
