+++ include/El/blas_like/level3/Syrk.hpp
@@ -1,0 +1,62 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_SYRK_HPP
+#define EL_BLAS3_SYRK_HPP
+
+namespace El {
+
+// Syrk
+// ====
+template<typename T>
+void Syrk
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const Matrix<T>& A, T beta, Matrix<T>& C,
+  bool conjugate=false );
+template<typename T>
+void Syrk
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const Matrix<T>& A, Matrix<T>& C,
+  bool conjugate=false );
+
+template<typename T>
+void Syrk
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const ElementalMatrix<T>& A, 
+  T beta,        ElementalMatrix<T>& C, bool conjugate=false );
+template<typename T>
+void Syrk
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const ElementalMatrix<T>& A, ElementalMatrix<T>& C,
+  bool conjugate=false );
+
+template<typename T>
+void Syrk
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const SparseMatrix<T>& A, 
+  T beta,        SparseMatrix<T>& C, bool conjugate=false );
+template<typename T>
+void Syrk
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const SparseMatrix<T>& A, 
+                 SparseMatrix<T>& C, bool conjugate=false );
+
+template<typename T>
+void Syrk
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const DistSparseMatrix<T>& A, 
+  T beta,        DistSparseMatrix<T>& C, bool conjugate=false );
+template<typename T>
+void Syrk
+( UpperOrLower uplo, Orientation orientation,
+  T alpha, const DistSparseMatrix<T>& A, 
+                 DistSparseMatrix<T>& C, bool conjugate=false );
+
+} // namespace El
+
+#endif // ifndef EL_BLAS3_SYRK_HPP
