+++ include/El/blas_like/level3/Trmm.hpp
@@ -1,0 +1,36 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_TRMM_HPP
+#define EL_BLAS3_TRMM_HPP
+
+namespace El {
+
+// Trmm
+// ====
+template<typename T>
+void Trmm
+( LeftOrRight side, UpperOrLower uplo,
+  Orientation orientation, UnitOrNonUnit diag,
+  T alpha, const Matrix<T>& A, Matrix<T>& B );
+template<typename T>
+void Trmm
+( LeftOrRight side, UpperOrLower uplo,
+  Orientation orientation, UnitOrNonUnit diag,
+  T alpha, const ElementalMatrix<T>& A, ElementalMatrix<T>& X );
+
+template<typename T>
+void LocalTrmm
+( LeftOrRight side, UpperOrLower uplo,
+  Orientation orientation, UnitOrNonUnit diag,
+  T alpha, const DistMatrix<T,STAR,STAR>& A, ElementalMatrix<T>& B );
+
+
+} // namespace El
+
+#endif // ifndef EL_BLAS3_TRMM_HPP
