+++ include/El/blas_like/level3/Trr2k.hpp
@@ -1,0 +1,17 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_TRR2K_HPP
+#define EL_BLAS3_TRR2K_HPP
+
+namespace El {
+
+
+} // namespace El
+
+#endif // ifndef EL_BLAS3_TRR2K_HPP
