+++ include/El/blas_like/level3/Trrk.hpp
@@ -1,0 +1,89 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_TRRK_HPP
+#define EL_BLAS3_TRRK_HPP
+
+namespace El {
+
+// Trrk
+// ====
+template<typename T>
+void Trrk
+( UpperOrLower uplo, 
+  Orientation orientA, Orientation orientB,
+  T alpha, const Matrix<T>& A, const Matrix<T>& B,
+  T beta,        Matrix<T>& C );
+template<typename T>
+void Trrk
+( UpperOrLower uplo, 
+  Orientation orientA, Orientation orientB,
+  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
+  T beta,        ElementalMatrix<T>& C );
+template<typename T>
+void LocalTrrk
+( UpperOrLower uplo,
+  T alpha, const DistMatrix<T,MC,  STAR>& A,
+           const DistMatrix<T,STAR,MR  >& B,
+  T beta,        DistMatrix<T,MC,  MR  >& C );
+template<typename T>
+void LocalTrrk
+( UpperOrLower uplo,
+  Orientation orientB,
+  T alpha, const DistMatrix<T,MC,STAR>& A,
+           const DistMatrix<T,MR,STAR>& B,
+  T beta,        DistMatrix<T>& C );
+template<typename T>
+void LocalTrrk
+( UpperOrLower uplo,
+  Orientation orientA,
+  T alpha, const DistMatrix<T,STAR,MC>& A,
+           const DistMatrix<T,STAR,MR>& B,
+  T beta,        DistMatrix<T,MC,  MR>& C );
+template<typename T>
+void LocalTrrk
+( UpperOrLower uplo,
+  Orientation orientA, Orientation orientB,
+  T alpha, const DistMatrix<T,STAR,MC  >& A,
+           const DistMatrix<T,MR,  STAR>& B,
+  T beta,        DistMatrix<T,MC,  MR  >& C );
+
+// Trr2k
+// =====
+/*
+template<typename T>
+void Trr2k
+( UpperOrLower uplo, 
+  Orientation orientA, Orientation orientB,
+  Orientation orientC, Orientation orientD,
+  T alpha, const Matrix<T>& A, const Matrix<T>& B,
+  T beta,  const Matrix<T>& C, const Matrix<T>& D,
+  Tgamma,        Matrix<T>& E );
+*/
+template<typename T>
+void Trr2k
+( UpperOrLower uplo,
+  Orientation orientA, Orientation orientB,
+  Orientation orientC, Orientation orientD,
+  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
+  T beta,  const ElementalMatrix<T>& C, const ElementalMatrix<T>& D,
+  T gamma,       ElementalMatrix<T>& E );
+
+// The distributions of the oriented matrices must match
+template<typename T>
+void LocalTrr2k
+( UpperOrLower uplo,
+  Orientation orientA, Orientation orientB,
+  Orientation orientC, Orientation orientD,
+  T alpha, const ElementalMatrix<T>& A, const ElementalMatrix<T>& B,
+  T beta,  const ElementalMatrix<T>& C, const ElementalMatrix<T>& D,
+  T gamma,       ElementalMatrix<T>& E );
+
+} // namespace El
+
+#endif // ifndef EL_BLAS3_TRRK_HPP
