+++ include/El/blas_like/level3/Trstrm.hpp
@@ -1,0 +1,35 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_TRSTRM_HPP
+#define EL_BLAS3_TRSTRM_HPP
+
+namespace El {
+
+template<typename F>
+void Trstrm
+( LeftOrRight side, UpperOrLower uplo,
+  Orientation orientation, UnitOrNonUnit diag,
+  F alpha, const Matrix<F>& A, Matrix<F>& X,
+  bool checkIfSingular=true );
+template<typename F>
+void Trstrm
+( LeftOrRight side, UpperOrLower uplo,
+  Orientation orientation, UnitOrNonUnit diag,
+  F alpha, const ElementalMatrix<F>& A, ElementalMatrix<F>& X,
+  bool checkIfSingular=true );
+template<typename F>
+void Trstrm
+( LeftOrRight side, UpperOrLower uplo,
+  Orientation orientation, UnitOrNonUnit diag,
+  F alpha, const DistMatrix<F,STAR,STAR>& A, DistMatrix<F,STAR,STAR>& X,
+  bool checkIfSingular=true );
+
+} // namespace El
+
+#endif // ifndef EL_BLAS3_TRSTRM_HPP
