+++ include/El/blas_like/level3/Trtrmm.hpp
@@ -1,0 +1,27 @@
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#ifndef EL_BLAS3_TRTRMM_HPP
+#define EL_BLAS3_TRTRMM_HPP
+
+namespace El {
+
+// Trtrmm
+// ======
+template<typename T>
+void Trtrmm( UpperOrLower uplo, Matrix<T>& A, bool conjugate=false );
+template<typename T>
+void Trtrmm
+( UpperOrLower uplo, ElementalMatrix<T>& A, bool conjugate=false );
+template<typename T>
+void Trtrmm
+( UpperOrLower uplo, DistMatrix<T,STAR,STAR>& A, bool conjugate=false );
+
+} // namespace El
+
+#endif // EL_BLAS3_TRTRMM_HPP
